+++ src/components/Password/index.ts
@@ -1,0 +1,3 @@
+import Password from './PasswordView';
+
+export default Password;
